+++ test/test_query.cpp
@@ -8327,6 +8327,25 @@ TEST(Query_MaximumSumAverage)
     }
 }
 
+
+TEST(Query_ReferDeltedLinkView)
+{
+    // This would segfault because the query refers a LinkView that had been removed by move_last_over
+    // It will now throw an exception instead.
+    Group group;
+    TableRef table = group.add_table("table");
+    table->add_column_link(type_LinkList, "children", *table);
+    table->add_empty_row();
+    LinkViewRef links = table->get_linklist(0, 0);
+    Query q = table->where(links);
+    TableView tv = q.find_all();
+    
+    table->move_last_over(0);
+
+    CHECK_THROW(q.find_all(), DeletedLinkView);
+    CHECK(!links->is_attached());
+    CHECK_THROW(tv.sync_if_needed(), DeletedLinkView);
+}
 #endif // TEST_QUERY